pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datacleanenv"
version = "0.1.0"
description = "Data Privacy & Compliance Steward: Audit and clean messy SQLite databases."
readme = "README.md"
authors = [
    { name = "Gargeya" }
]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "pytest",
    "requests",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"
